+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Relayer
 
+* [\#487](https://github.com/cosmos/relayer/pull/487) Add override flag for client reuse.
 * [\#491](https://github.com/cosmos/relayer/pull/491) Add flag for client parameters which allow or disallow governance to update frozen/expired clients
 * [\#486](https://github.com/cosmos/relayer/pull/486) Refactor fly path generate command
 * [\#475](https://github.com/cosmos/relayer/pull/475) remove cmd docs file